requirements.txt
# requirements.txt

flask
python-dotenv
Flask-Security-Too[fsqla,common]
